Introduce SanDecoder
Close #536
